+++ packages/ricos-content/web/src/normalization/normalizeInitialState.spec.ts
@@ -16,6 +16,8 @@ import {
   textWrapContentStateExpected,
   noConfigContentState,
   noConfigContentStateExpected,
+  invalidInlineStyleState,
+  invalidInlineStyleStateExpected,
 } from './Fixtures';
 import { WRAP } from '..';
 import { compare } from '../comparision/compare';
@@ -909,6 +911,16 @@ describe('normalizeInitialState', () => {
       );
       expect(actual).toHaveProperty('ID');
       expect(actual.ID).toStrictEqual('1234');
+    });
+  });
+
+  describe('invalid inline style removal', () => {
+    it('should remove number from inline style', () => {
+      expect(
+        compare(normalizeInitialState(invalidInlineStyleState), invalidInlineStyleStateExpected, {
+          ignoredKeys: ['ID'],
+        })
+      ).toStrictEqual({});
     });
   });
 });